RTS-3791 Remove link for third item in breadcrumb
The third item in our breadcrumb is not an active link and gets a href-
attribute of '#'. That the item is clickable may be confusing and we wish
to remove the markup that identifies it as a link. When building the
breadcrumb the items in the active_trail has to be added as link items,
but we may remove the a-tag from the output using theme_breadcrumb()