RTS-#1448 Setting document preview file display as large filetype icon
Due to a problem in the media browser, using "image" as display causes
a javascript problem in the media module, which the large filetype
icon avoids.

See: https://drupal.org/node/1743040#comment-6704122 and further
comments.

This commit sets the widget type for the attachments (field_uib_files)
in articles, and the document file type preview file display to
large filetype icon.