RTS-#4045 Increasing page width
The page width is increasing from 910 to 1260px.

Some problems with our css were immediately noticed:

* The slideshow does not scale correctly. It it expected to behave
better when rts #3927 is in.

* The background image for the second sidebar was set to a fixed
position on the main div. Because of this, the background image did
not follow the second sidebar. The background image was moved to
the latter.

This has the potential drawback that there is no background image
below the end of the sidebar contents. It has the positive effect
that there is no background if there is no sidebar.

* The second level menu items had a fixed size, and did not scale
correctly. They were set to be 20% wide to correspond to the
width of their parents.