RTS-4185 Added test before adding hyphen to name in testimonial teaser
We add a hyphen before the name in the testimonial teaser. This makes the
hyphen and the read-more-link appear above the qoute when the name is missing.

Testing if name is empty before adding the hyphen makes the link appear below
the picture and quote as it is supposed to do.