RTS-#3742 Remove unnecessary relation from my content view
This unused relation caused that some users were not allowed access
to external content. The real cause is not fully known, but needs to
resolved in #3743. However, for this issue it was enough to remove
the unused relation.

Also adds a user permission to allow users the same as for area and
article.