RTS-#4310 Feed aliases for employee nodes
The automatic alias setup for nodes specifically excluded the
employee pages. However, the two feed aliases should be available
for the employee pages too.

Previously, the list of aliases was duplicated in two separate
functions and - as might be expected - they had diverged. For the
purpose of consistency, the list of aliases was moved into a
helping function of its own.