Don't restore field revision data from production
It's not that useful in development instances and this eliminates a lot
of data.  Today the about 70% of the 255 MB database dump files consist
of field revision data, and these tables just continues to grow.

A consequence of this hack is that if you try to revert to an earlier
version of a node you get blank values for the fields.

It would also be usefull to trim down the node_revisions table, but
that will have to wait for another time.  Node display will suffer it
we removed all data from that table.