RTS-#4290 Text filter linking study codes to their content
Text filter that scans for study codes and adds links to their
content at "emne/CODE". The study codes have to be written in the
correct case and be separate words. If they are only separated
from other words by comma or period, and no whitespace, the filter
will fail.

The filter is setup for text formats "Basic HTML" and "Restricted
HTML".