RTS-#1974 Adding an study entityreference field to uib_testimonial
In order to keep existing information - which is now in the field
field_uib_related_programmes - a new field is created:
field_uib_study_programmes. The new field is an entity reference
field connected to uib_study. The autocomplete is performed by
using an entity reference view.

The commit also contains an update task (uib_testimonial) that
will pick up pre-existing information and attempt to find the
correct entity for the new field.

Note that the old field is not removed here. This should be done
after the update task has been properly run, reasonably in a
follow-up.