RTS-#1974 Temporary styling for testimonial node edit form
The reason for this styling - which hides the old
field_uib_related_programmes from the testimonial node edit form
is that the new field has to be filled properly by the update
task before the old field may be removed. So we just temporarily
hide the old field from the form.

The follow-up issue that removes the field should revert this commit
also as a styling cleanup.