RTS-#4526 Adding formatter for displaying entityreference field
When displaying an entityreference field, the node title is shown by
default. For uib_study, we use a field instead of the node title,
as we use entity translation. This formatter is added to be able to
display a field instead of node title.

In order to be able to reuse the filter for additional similar
cases, settings were added (field name, link or not).

At this time, the filter was placed in the uib_study module. If it
is used in other cases, it might be more convenient to move into
a separate module.