RTS-#4526 Update testimonial language inheritance
Ca 40% of the current testimonials lack specified language (i.e.
it is UNDefined) - probably for historical reasons. As this will
cause problems in displaying the correct translation of related
Study nodes, an update task was added.

The task updates the language for all existing testimonials to
be the same as that of the (first) area they are connected to.