RTS-#4526 Setting language of testimonial upon save
When created, a testimonial might become language neutral and this
would affect the display of the translated node title field in an
undesired fashion.

Ideally, the language of the testimonial should reflect the language
of the main area to which it is connected (not the language of the
user that created the testimonial).

This commit fixes the language of the testimonial to be the same as
its area, upon node save.