RTS-#4308 Unpublish study nodes when not in use
Study nodes are unpublished rather than deleted when they do not
appear in sync-fs. There were a number of query changes in order
to facilitate and speed up the sync. Unfortunately, unpublishing
a node seemed to require loading and saving it.

Some changes were simplifications after switching from two to one
study node per course/program, and a format change of the node
title.

Also, update of the study title field was added, and a fix to the
EntityFieldQuery which did not find unpublished study nodes
unless run as the privileged user.