RTS-#4640 Remove ripped-out content from main study content data
Some of the study data (notably contact info, teaching language,
and number of study places) are copied and displayed in separate
blocks: Study facts, and contact info.

This commit is intended to introduce functionality that removes
(all instances of) the copied data from the study data - in order
to not display them more than once.

Some changes were necessary in the study_contact block.