RTS-#4707 Fix missing authors and byline in article (and testimonial)
Users not added during migration (but through Sebra sync) will eventually
be lost as authors (or byline authors) in more recently migrated
articles and testimonials. The reason is that migrate tries a user
lookup in its own tables, but fails.

This commit adds a post-upgrade script that attempts to update author
and byline for existing articles and testimonials - where missing -
by fetching user data from w2. It introduces a minor addition
(UiB user name) to the XTopic object.