RTS-#4680 Menu tab on testimonials
A menu tab is added to the node form of the uib_testimonial content
type. It reuses the code of the uib_article menu tab, which was
changed to be possible to call from the uib_testimonial local task.

Since the local tasks (tabs) are setup by using hook_menu(), the
articles still use the path "node/%node/menu" while the testimonials
had to use another path "node/%node/tmenu".