RTS-3902 Make addresses of place show in colophon
Until now the only address for an area displayed in the colophon
has been the global hardcoded address.

This code makes use of the ou of an area and fetches the address
for an area from the ou. If an area does not have an ou, the
parent area is used to get an address.