RTS-4145 Made h2 printable when in uib_h2tabs
Since the introduction of uib_h2tabs h2 in the main text area has not
been printable. The headlines has been printed as part of the tabs,
but have been removed from their place in the text.

This commit leaves the h2s in the text and hides them from being
displayed on screen, and hides the tab navigation and displays the
h2s in print.

It also cleans up a bit of misplaced styling in normalize.scss.