RTS-#4964 Missing translation text on uib_study
Since study nodes are undefined regarding language, we need to
check (1) if there is any translated content, and if there is
not, (2) if there is any content at all, before displaying a
notice.

Also use drupal_set_message for not-translated notice

Instead of using our own message box that adds html markup and
styling, we switch to using the drupal_set_message function.