RTS-4996 Setup aliases for NUS taxonomy terms
In order to be able to use the Norwegian and English names of the
education area as first part of the alias, it did not seem feasible
to set up a path pattern. Instead, the taxonomy_term hooks for
inserting, updating and deleting terms were used to add and remove
the aliases.
   The names of the education areas are picked up by loading the
aliases of those nodes.
   The aliases are lower cased, stripped of non-word characters and
spaces are replaced by hyphens.