RTS-2618 View the ties that an area has to nodes etc
Adds a local task tab "Ties" on areas and a hook to wipe the slate clean
when an area is removed.

The "Ties"-tab can be accessed by users that have the 'delete any area
content' permission.

The tab page lists menu links, aliases, redirects, nodes of which
the area is a parent, and content nodes (articles, testimonials
and external content that are connected to the area through field_uib_area).

For connected content, it is noted whether the nodes are owned by the area
(i.e. being the first listed area in field_uib_area), and if the node has
been shared by others. Also listed are nodes that relate to the connected
content node via "related content" (field_uib_relation).

The added hook_node_delete() removes menu links pointing to the area
node, aliases, redirects, content (via field_uib_area) owned by the
area.

Also, related content indications are removed, and share indications
"bobbling". If there are any area children, they are if possible,
connected to the parent of the area being deleted.