RTS-4998 Set the access to the taxonomy term relationship
Hiding the relation form field in the taxonomy term form by using
the hide() function resulted in reset of the term relationship to
the vocabulary root on every save. Setting no access, instead,
worked as planned: The user does not see the field in the form,
leaving the term relationship in the tree unchanged.