RTS-5290 Properly handle event dates in migration
We used a manual handling of date/times from w2 in the prepare
function of event migration. This is not necessary since the
migration integration of the date module can deal with complex
destination fields. Rather, the manual handling caused the end
date to be off by the difference between GMT and local time if
end date was missing in w2.

The solution was to correct the fieldmapping and remove the
manual date handling.