RTS-5250 Using field collection instead of links in uib_study
The section for relevant links in uib_study was set to field_uib_links
which is a list of links. What was wanted was a possibility to add
any number of link lists with their own title. To achieve this we remove
field_uib_links and add field_uib_link_section.