RTS-5333 Translating exposed filter options in study programmes view
We found no way to arrange translation of the selectable options
for the exposed filter. Thus, the solution here is the same as the
one previously done for the calendar view: To pick up the text from
the view information in the hook_views_post_render(), translate
them using the i18n_string_translate() function and replace the
texts in the rendered views output.