RTS-5149 All social media icons should have the same style
Until now the social media icons used in areas and articles have been
of different size and origin. This unifys the use of social media icons
througout the site and all icons is one sprite. The service links module
also uses the sprite.

This commit also introduces a mixin to resize a sprite to use in various
sizes.