RTS-5446 Get path for study node, not alias
To generate the url for a spesialization made bad things happen with
links in construction table. It may not be a problem once the table is
trimmed down to the correct size, but using the uib_study__study_path
function to get the correct path is a more robust way anyway.