RTS-5535 Made the field formatter in uib_study use global language
The uib_study_field_formatter_view used the entity language to
determine the language of the text to be formatted. The language code
of a study node is always 'nb' which caused the text to always appear
in norwegian. Making the formatter use the global language code made
the text appear in the correct language.