RTS-5167 Add field_multiple_limit module & settings for uib_testimonial
The field_multiple_limit module makes it possible to restrict what to
display from any multi instance field, and can be set via the display
settings in the content type.

This commit modifies drupal.make to add the module and the
field_formatter_settings module which it needs, and adds the
settings for the teaser and short teaser in uib_testimonial to
show only one image. Note that since the module adds the default
setting for all multiple instance fields, several modules are
affected.

The feature settings for fields in uib_testimonial was for some
reason present in the info file of the uib_area module. The commit
moves the relevant field bases, and field instances to the uib_
testimonial module.