RTS-5494 Hide info type headings when the info type has no heading
The info types may have content for previous, current and next semester.
If content for one semester is present the heading of the info type
will be displayed for all three semesters leading to a stand alone
heading being displayed when a info type is empty for the selected
semester, but not empty for on of the other choices.

By moving the heading into the div that contains the content the
visibility of the heading will be decided by the status of the content.