RTS-5572 Map urls for study plans of specializations
The link for study plan in a specialization takes the user to a
non-existing page. This is partly because the namespace for the study
plan is not mapped in uib_study_menu, partly because the url generated
for the link is not adapted for specializations.

By adding mapping for the namespace of a study plan of a specialization
and alter the generation of the study link, the issues no longer exists.