RTS-5563 Sorting on "group by" study category in study related views
We can not setup a view to sort by translated study category due
to (1) the field is a select list, translated by string translation,
and (2) views does not sort the group sets internally - it merely
picks up the correctly translated [study category] strings.

The suggested solution was to re-sort the results in an imple-
mentation of the hook_views_pre_render(), and depend upon
the views algorithm (in views_plugin_style::render_grouping) which
divides the results in sets. Currently, this concerns the
study_programmes display in the courses view, and the
uib_taxonomy_term view.

The commit also adjusts the usage of field_uib_study_title to
filter language versions in the same way for all relevant
displays in the courses view.