RTS-5387 Trim contruction table for specializations
On study specialization pages, only show data for the current
specialization in the construction table.

All data for the entire study programme table is received from
fspres. The solution was to take advantage of an already present
initial scan through the data and add a check for what speciali-
zation data that are irrelevant - and then remove it before
the table is built.

In addition, an initial descrition text is removed. That text was
relevant in the study programme scope, but not when showing only
one specialization. Also, linking the specialization heade to
itself is avoided.