RTS-5572 Make page callback aware of specializations
The final problem in this issue was making the page callback aware of
specializations. The page callback were only considering the first argument
of the url and thus always got the study plan of the masters program, not
the specialization's. By adding the possibilty of a second argument we are
able to get the correct study plan.