RTS-5807 Force migrated content to be published
The default state for created Articles was changed to unpublished in
commit fed3ad9 (RTS-5351).  This ended up also affecting migrated content
as migrate explictly calls node_object_prepare().  We compensate by setting
the status back to published in the prepare callback.

This change also adds the override to the Testimonal and Area migration
in case the defaults for them change in the future.