RTS-5818 Change Google Analytics page filtering
The Google Analytics module is set up with a list of the pages
where tracking is NOT to occur. Previously, "node/*/*" was listed.
Thus, views pages like node/12345/calendar would not be tracked.
The basic idea initially should have been to avoid tracking pages
like node/12345/edit etc.

In order to change the tracking according to the issue, the
"node/*/*" was removed, and instead pages with similar path
(such as node/12345/edit) were specifically listed. This solution
is less optimal than before, because new pages that are added
(such as new "tabs") will be tracked by default unless added
to the Google Analytics module page filter.