RTS-5487 Abort uib-sync-fs on first error
If we get one failure it's likely that all requests will fail.  Might as well
give up right away.

I also tried to run 'drush uib-sync-fs' against a server that reported 404
on all requests.  We ended up unpublishing all studies.

This change adds a limit on how many studies must be found before we
start unpublishing those that were not reported by fs-pres.

The trigger limit was set to 2200.  There are currently 2374 studies
in our database.  Hopefully that's enough margin so that we will not
have to adjust this limit.