RTS-5839 Fix dates in event migration
Datetimes for events become converted during migration. We want
to keep the UTC dates coming from w2, but it seems that some
software component assumes that the date is local timezone, and
converts it into UTC. The reason is not entirely clear.

The fix is to handle event datetimes manually during migration.