RTS-5839 Add Zulu (GMT) to dateTime fields from w2 when migrating
All migrated dateTime fields have been offset with the difference
between timezones UTC and Europe/Oslo, not just the event dates.

The reason was that the php conversions between date strings and
unix timestamps assumed that the strings were local time when
no contrary information was present - which is the case for all
dateTime fields from w2.

The solution was to add an indication that the dateTimes are UTC
in the XTopic object construct. We used "Z" (Zulu), an age old
standard.