RTS-5784 Translate all buttons in webdesk
Our menus are not translated, we serve a different menu for each area.
In the webdesk we had one menu and it would not be translated when
switching language. To make this correct we needed to make a menu for
each language in the webdesk and swap menus based on language in panels.