RTS-5903 Setup entity translation for users
Turns on entity translation for the user entities.
Even though one field has translated information and the field in
itself has been setup as translated, there is no translation
handler for the user. This would cause the english version of the
field_uib_position to be copied over the norwegian version as
soon as a translation was started for a user. Thus, this commit
adds an update task that fixes this for all existing users
(except user 0 and user 1).