RTS-5904 user data migration script
In order to migrate user data from w2, this commit adds a separate
migration script. This avoids the migrate module, but uses the
XTopic object construct from the uib_migrate module.

The reason for migrating these data separately, and not via a
migrate --update is that in our setup, some users have been synced
from Sebra. As migrate has no memory of these, it would (or might)
cause problems that this solution avoids.

The w2 fields migrated are:
The norweigan and english versions of arbeidsfelt-2, forskningsfelt,
fagomrade_cv, prosjekter, publikasjoner, undervisning-2, and funksjon.
And: hjemmeside, lenke (if there is no 'hjemmeside'),
alternate 'besoksadresse', romnummer, rssurl, rss-tittel.