RTS-5557 Move view results below exposed filters
The straight-forward way of showing the results of a view page
query is to add it to the header or the footer of the view. However,
the standard view page template displays the exposed filters below
the header.

For the purpose of the courses page views, this commit introduces
a templace which is a copy of the standard views page template
except that it puts the exposed filters above the header. A wee bit
of styling is applied.