RTS-5394 set group titles in views lists to be h1
The header for group titles in the default template
views-view-list.tpl.php is h3. This was changed to h1 for taxonomy
terms (in a specific template). This commit makes it more general
by changing the name of the template previously setup for taxonomy
terms only.

This will now affect all group titles in views that are setup to
display html lists (i.e. <li>). A very minor change was made in
css to acommodate the change.