RTS-6004 Making specializations appear in study programme listings
In study programme listings in w2 specializations is listed related
to the study programme the specialization is a part of. To achieve
this with views in drupal we need to add the module views_field_view
and create a sub view, before adding a view as a field to the original
view.

Imports views_field_view-7.x-1.1