RTS-6063 Actually display title in page template
Previous work on the calendar removed the display of title in
favour of content_top, from the page template. Instead, titles
were incorporated via code in template.php.

This commit returns the printing of title to the page template, and
adjusts the theme_preprocess_* hooks where necessary.

In particular, the title for the uib_taxonomy_term view needed to
be properly set in the header, but removed from the page since it
was also displayed in content.