RTS-6048 Temporary warning message(s) on w3 user profile
Displays a message using drupal_set_message() for the cases listed
below, when a user views a w3 user page, to indicate that this is
still work in progress. There are three messages of
different specificity. (1) a general message, (2) a message with a
login link to w2 user profile login, and (3) a message with an
additional plea to the user not to edit the w3 user profile yet.

Cases:                                      Message:
Anonymous user viewing a user page             1
Logged-in user viewing other user's page       1
Logged-in user viewing own page                2
Logged-in user editing own page                3

This commit also fixes a vexing undefined notice from 10 lines
above (code not checking that variable was set before accessing it).