RTS-6271 Created the uib_dbh module
This module interfaces with the web service of NSD DBH that provide metrics
and graphs about the performance and trends at UiB.

This module provide a block called 'uib_dbh' that is inlined on the about
pages and implementation of graphs pages at 'about/nsd/%'.  The links
on these pages go to NSD.

Since the data from the web service only changes once every year we'll
just keep a local copy of the JSON file.  The pages.json file need to be
refreshed from http://dbh.nsd.uib.no/dbh-service/listPages.action?uid=zxgcxkfq on
request.