RTS-6380 Set up the webdesk/stats page
I just wanted a place to present live vital status data about our
content, so I made this page.  It's meant for internal consumption,
but there should be anything here that we need to keep secret,
so the page is open for everybody to read.

Feel free to add other tables to this page.