RTS-6372 Add new geofield field for uib_ou
A so far unknown problem with the old geofield causes db errors when
the existing field_uib_geo is updated. A new field setup in the same
way does not have these problems.

The proposed solution is to create a new field, move the data, and
change the view that uses the field. This commit adds the new field
field_uib_geolocation.