RTS-6415 Adding research groups block to user profile page
The user profile has no information about area connections, so this
commit creates a block that attempts to find connections to the
user in field_uib_staff for research groups [areas].

The block is placed by using a new context "user_profile" and by
modifying the weights according to the layout sketch order (in
issue 5896).

The _preprocess_page hook was used to position the block in the
"group_user_second" field group.