RTS-6412 Disable synchronized fields on user edit form
Sets the #disabled property to TRUE for a number of fields and user
entity properties - in the user edit form. They are set to disabled
regardless if they are visible or not (since they may be made
visible from the gui).

Two user properties are not set to #disabled even though they are
synchronized: status and roles. Level 1 wants to be able to edit
these.