RTS-6433 Setup uib-tabs for certain fields on user profile page
Issue RTS-5896 notes five tabs: Research, Communication, Teaching,
Publications, and a tab with optional name. In RTS-5903 the fields
were created and placed in the field group group_user_work_1.

This issue adds the previously used jQuery necessary for the tabs
and adds markup.
It also handles the optional tab. The latter will not be shown if
the optional title is missing - even if there is data. Without a
title a tab is not possible.