RTS-6420 Styling social media icons on user profile page
The style should approach the social media icons on front page,
but without the text. The style did not automatically apply since
the field names were different.

Thus, a new class "uib-social-media" was setup for both fields, and
the relevant styles updated. Then, some additional styling was
applied for the user page.