RTS-6520 Add check that user has access to first area in article
Whan a user adds or edits an article, a problem occurs if the
article is "owned" by an area that the user is not content
manager of: The user will loose the ability to edit the article.

This commit attempts to alleviate the problem by stopping a user
from saving an article with such an area connection. This only
applies to the first element in field_uib_area, since that is
what determines area "ownership".

User 1 and users with the "Level 1" role bypass this check.