RTS-6400 Made mobile submenus expand on android devices
When a user on an android devices tried to expand our menu items in
our mobile menu the user were taken directly to the url in the menu
item clicked. To avoid this we needed to tweak the javascript handling
the click events on touch devices. We needed to target the correct
elements.