RTS-6568 New module uib_cristin
Setting up a new module for the integration with Cristin.

The purpose was to add a simple link from the user's profile page
to the corresponding Cristin page. In order to do this, a framework
for Cristin integration was setup in its own module.

These cases are handled:
1. The publications field contains text and there are publications in
Cristin: The link prepends the existing text.
2. The publications field is empty, but there are publications in
Cristin: The link is displayed in the Publications tab.
3. No publications in field, nor in Cristin: No link, no tab.

In order to do this, a simple query is performed using the Cristin
web service, to check if there are any publications available
for the user.