RTS-6567 Not printing page title on user profile pages
In page.tpl.php there is a fair amount of code dealing with page
titles. Except for node pages, it is printed in main content.

In order to avoid printing the title on user profile pages, a
boolian variable is introduced in preprocess_page (template.php)
indicating if the page is a user profile page.

The test of this variable is added to the already existing test in
the page.tpl.php template file.