RTS-6567 User page user url formatting and styling
Set the url field to display no label, but url title and link.
In order to let the url link title to display in the same way as
a field title, a colon was added in css after the title.

The commit also removes the bullet on the ou li.