RTS-6692 Study nodes as related content displays correct title
The title of our study nodes are not translated, we use a translated
title field to display the correct language. We need to swap the title
with the title field of the current language.

The commit also makes the __uib_title function work again and removes
a div surplus in the markup.