RTS-6695 Experimental EVU course overview page
This is just an experiment to highlight what the issues with this
page is.  Currently all the requested information is present, but
not in the form of a table for narrow screen friendlyness.

All the inline style will also have to go.  Style should probably
be shared with a similar "Sign up" section on the course page.