RTS-6765 Support banner resources
The banner field might now start out with a line of the form:

  <!-- Resources: <res1>, <res2>, ... -->

Where the resources can be .css or .js paths or the string
'uib.tabs' to load our tabs implementation.  More libraries
can be added easily in the future.

Pages where the banner shows will load these resources ahead
of the banner code itself.