RTS-6778 Moved offices of an area into the above_content region
To move the offices the content were made into a block in the area
module and added to the region in context. The code presenting the
offices as tabs using jquery ui tabs were also rewritten in the area
module and removed from template.php. Minor adjustments to the styling
were required.