RTS-6908 Change uib_study texts
Changed the text "Study points" to "ECTS credits" in three places:
In the "study_content" block, and twice in the "study_facts" block
(once for course, and once for programme/specialisation).

The two second are the ones mentioned in the issue summary, and the
first relates to the html title for the abbreviated column title
"SP" in the "construction" table.