RTS-6951 Make sure array is not passed as id in user sync
The job uib-sync-w2-user has been crashing due to an array being
passed as an id. The job has exited due to a wrong type parameter.
In an article the main media may now consist of any number of images.
The same property in a XTopic object is used for the user image, but
no adjustments were made to reflect that the property is an array.