RTS-6690 Create description (emnebeskrivelse) page for course
Defines a new page for course/emne: course/<code>/description and
harmonized styling with the study plan page.

To create the page, we use the uib_study__studyplan function,
modified. Translated study title now works.

The page will always display the text for current semester. The
reason for this is that the uib_study__is_course() fails, since no
node is loaded.