RTS-6691 EVU course pages: hide description and add link instead
This is implemented by hiding the description (emnebeskrivelse) in
template.php (preprocess node), and adding an additional resource
link for evu-courses in the study_facts block.

In order to conform with previous code in the study_facts block,
the link is provided by using a separate utility function
uib_study__get_description_link().