RTS-4671 Handle alias updates for areas and articles
Fixes problems which in some cases caused duplicate aliases for
areas. Adds functionality for area alias changes.

When an alias for an area is changed, all articles and views connected
to that area (via field_uib_area) will get their aliases updated too
according to the area alias change.

Introduces a variant of the path_load function (uibx_path_load) that
will load all (if there are more than one) paths for a source, and
orders them with the newest (highest pid) at the start of the list.