RTS-7003 Use uibx_html_linkify to implement uib_study__filter_text
Avoids exponential growth in run time.  This version of the function is
about 6000x times faster when processing an 1 MB document.

There are still 2 study codes[1] that don't match the provided regular
expression, but I just think those are so stupid that they don't deserve
automatic linking.  We'll revise if we get complaints.

[1] "ODO-SPES.EKS" and "VIT411-7,5"