RTS-6399 Styling of name etc in first column of user edit form
Here is a proposed styling of the 6 user fields placed in the first
(left) column, mentined in the issue.

They are not easily positioned using the grid, since the left
column consists of three grid columns, and the suggested layout for
these fields splits the column in half. Because of this, the column
split is not done by using the grid.