RTS-7050 Evu course price text display
The incoming price ("kostnad") for evu courses is an integer. This
commit displays "No course fee" if the price is === 0. However,
this has not been possible to really test, because currently all
evu courses either does not have the field set (resulting in
displaying "Course fee not set") or a price > 0.