RTS-7048 Add to and cleanup header on description and plan pages
Adds an extension to the get current semester function, with a
longer format.

Joins this with the previous header into a single header, and hides
the page (html) title because it may not (in Drupal?) contain an
anchor link. The html title is still there (browser border) but
not shown in content.

There was previously some (non node) pages where title was hidden,
and that functionality was modified to be more general.