Added the 'umod' script
This Perl script analyzes the source code of the UiB specific modules
and can report on stuff like where things are defined and where they are
referenced.  The script doesn't really parse the code.  It relies on
conventions and that the code follows our style guide to extract the
required information on our code.

I wrote this as I tried to understand what the dependencies between
our modules really are.