RTS-6975 Placement of fields in contact section of user profile
Because the visit and postal addresses belongs to the OU, they
were rendered in the same div as the OU name, Thus, it was
difficult to place them independently in the contact section of
the user profile page. This was solved by moving those fields
in hook_user_view_alter().

To facilitate formatting, two field groups were added for those
fields that should not use the whole width of the contact section.
Some corresponding sass code was modified.

Note that this commit also include the minor change needed for
issue #6958.