RTS-7147 Using bundler to manage gem dependencies
To make sure the gems used in the project are consistent in all
stages of the development we should start using Bundler to manage the
dependencies. This way every developer and server should use the same
version of a ruby gem.

This commit adds a Gemfile that specifies the dependencies, updates
scripts using compass and the documentation regarding compass.

To autoinstall dependencies bundle install is added to the scripts
used to install and reset the site. The bundle config file is added
to git to avoid parameters and the documentation is tweaked.