RTS-7029 Copy user room and url content to language instances
Previous content for fields is saved as undefined language "und".
When the field is set to translatable, the und-instance will
interfere with translation unless the setup process in the gui is
used. The result is that all data in the fields are seemingly lost.

Thus, this commit contains an update task that removes the "und"
and copies the data to "en" and "nb" language instances for these
two fields in users.