RTS-7027 Collapsible current projects field on user profile page
The field is displayed in a ctools collapsible container which is open
or closed depending on amount of text.  Effort is made to only count
text (and not markup) in the limit of 200 characters for closing the
section by default.

Visual clutter was avoided by using the field label (title) as
the clickable line for inflating and deflating the field instead of
the proposed "read more" text.