RTS-7030 Modifications to uib_ou short teaser
In order to show a translated ou name and to be able to manipulate
the node link from the user's view point, the translated ou name
field (field_uib_ou_title) was added to the ou short teaser. The
node title was hidden.

Previously, all short teasers were setup to use the node--children
template. Due to this, the ou on the user profile was enclosed in
an irrelevant <li> and a title link created. This commit fixes the
incorrect template suggestion for ou short teasers, and removes
corresponding styling which is now no longer necessary.