RTS-7030 Select an area to link to at user profile page
A user is connected to an ou but it is expected that a link to
a users departmental home should be to an area, since the ou node
contains little relevant information. As an ou and a user may be
connected to multiple areas, the following algorithm is used to
determine the user's "home area":

1. Among areas connected to the ou, look for one that is "institute",
   "faculty" or "unit". If one is found, it is used for linking.
2. If the above fails, look for an area of the same types among those
   to which the user is connected as staff.
3. If the above fails, the ou name gets no link.

The commit also contains a minor fix for a function in the uib_user
module that had a potentially problematical and unnecessary
dependency on the uib_zen theme.