RTS-7168 Replacing line-height with padding in area menu
By replacing line-height with padding in the a-element in the area menu
the collapsing height issue disappears.

Added margin below expanded level 3 menu items for mobile.  On small
screen devices the margin below the expanded level 3 menu items is zero.
Increasing the margin will make the expanded section of the menu more
defined.