RTS-6713 Add a language switcher to the node edit pages
The language switcher is not displayed on node edit pages because
it is not rendered in our page template. This is controlled by
the "uib_node_edit_mode" variable setup in template.php.

This commit moves the language switcher setup in template.php
to "above_content" on the node edit pages, then adds a bit of
styling to make it display to the right.