RTS-7181 Override url link title when empty on user profile page
The link module copies the url to the title if the link title
is empty. This is hard coded as a part of the "_link_sanitize"
function of the link module.

In order to override this behaviour (field_uib_user_url) we can
override the theme_link_formatter_link_separate() function.
The problem with this is that the link title already has been
cleaned up at this point and as we do not have access to the
original field data, we can not be sure that the title was empty.
If the user had purposefully copied the url into the title, it
would look exactly the same. However, in both these cases we
will now use "Home page" as displayed title.

Since the label will be truncated if the URL is long, we just
replace it with the default title if it looks like an URL.